chore: reorganize entries
