+++ src/index/index.react-server.ts
@@ -1,2 +1,2 @@
-export { unstable_serialize } from '../core/serialize'
+export { unstable_serialize } from './serialize'
 export { SWRConfig } from './config'
